requirements.txt

python-telegram-bot==13.7
requests==2.25.1
web3==5.25.0
